Extract S3WriteFile and S3ReadFile classes into separate files
